refactor: 중복되는 로직을 유틸함수로 분리하여 제거 (#102)
